docs: Added "In Work" to top of report
I am doing most of my writing in Word and then transferring the information over to my IPython notebook on GitHub.  I want anyone reading the notebook to know that the work is still rough draft and I will let everyone know when the work is completed from my end.